[Silver III] Title: 주유소, Time: 380 ms, Memory: 36284 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -14,21 +14,21 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine());
 
-        int[] distanceArr = new int[N-1];
+        long[] distanceArr = new long[N-1];
 
-        int[] oilPriceArr = new int[N];
+        long[] oilPriceArr = new long[N];
 
         StringTokenizer st = new StringTokenizer(br.readLine());  
 
-        int sum = 0;
+        long sum = 0;
 
-        int oilPrice;
-        int nextDistance = 0;
+        long oilPrice;
+        long nextDistance = 0;
 
         for(int i=0; i<N-1; i++){
             distanceArr[i] = Integer.parseInt(st.nextToken());
         }
-        int totalDistance = Arrays.stream(distanceArr).sum();
+        long totalDistance = Arrays.stream(distanceArr).sum();
 
         st = new StringTokenizer(br.readLine());
         for(int i=0; i<N; i++){
@@ -44,7 +44,7 @@ public class Main {
             else{
 
                 for(int j= i+1; j<N; j++){
-                    int nextPrice = oilPriceArr[j];
+                    long nextPrice = oilPriceArr[j];
                         nextDistance += distanceArr[j-1];
                         if(oilPrice>nextPrice && j != oilPriceArr.length-1){
                             sum = sum + (oilPrice * nextDistance);
